minor #330 [Tests] Refactor WebTestCases in order to hold a client instance (ogizanagi)
This PR was merged into the master branch.

Discussion
----------

[Tests] Refactor WebTestCases in order to hold a client instance

…and simplify code

WebTestCases almost always required a client instance. Refactor AbstractTestCase in order to hold a client for each test and ease making requests.

Note that it will have some drawbacks if the bundle needs one day to load a lot of different configurations using environments for creating a client, as for now, a default client is always created using the `test` environment. Except in `AdminControllerTest` where we re-init the client with another environment, which means recreating a kernel and booting it...

The best solution would be to have tests classes for each used environments, and init the default client accordingly. (For instance, `AdminControllerTest` should became `AdminControllerEmptyTest`, or should be in a different folder, and the default options for creating the client will be `array('environment' => 'empty')`).

Commits
-------

b8a1c5f [Tests] Refactor WebTestCases in order to hold a client instance and simplify code