Security is a requirement for an admin
Not sure about this one. But without the security bundle, the admin does not work as the layout uses `app.user`, which is only available when the security bundle is available.

Two possible ways to fix it: trying to make usage of `app.user` optional, or add security bundle as a requirement. I think the second option, which is what this PR do, seems legitimate as not-securing an admin does not seem like a use case we should support.