Polish translations for exception messages
An "obiekt" ("entity") translation was used because "pozycja" ("item") in Polish is used for list elements mainly. Technically,"entity" could be translated as "encja" but 1) it is calque of English "entity", 2) in English translation, it is explained as "item" - perhaps because it's easier to understand by people without an IT background. Meanwhile "obiekt" simultaneously has both common-world and IT-specific meaning.