Added form tabs
Like the form groups tabs are another way to separate form fields to
achieve a better overview.

Tabs can contain groups but not the other way round. The
`NormalizerConfigPass` ensures that there are no orphan fields before
the first tab definition similar to the normalizing of the groups'
config and it even takes care about it if tabs and groups are used
together.

In the Twig template the part, that is responsible for the
display of groups, has been moved to a separate block to avoid code duplication.