JSON encoding nested arrays avoids Twig Exception
Rendering an Entity with an array element which contains a nested array or object was causing a fatal Twig Exception:

> An exception has been thrown during the rendering of a template ("Notice: Array to string conversion")
> Uncaught PHP Exception Twig_Error_Runtime: "An exception has been thrown during the rendering of a template ("Notice: Array to string conversion")." at home/ubuntu/Public/webapp/vendor/easycorp/easyadmin-bundle/src/Resources/views/default/field_array.html.twig line 5

I added a bit to check if the value of an array is also an array or object, and json encode it if so.

This seems like a reasonable balance between infinite nesting hell and just dumping serialized and practically unreadable data, but it's just a 1 line fix to prevent a nasty error for now.

I understand storing nested arrays isn't a particularly common (or good) idea, but it is technically correct and properly handled by Doctrine. This was the least disruptive and opinionated way I could think of to fix this, but I think it might also be a good idea to colorize the json encoded strings or otherwise set them apart so it is clear they are actually objects. Probably nested `<ul>`s or reliance on Twig's new `__toString()` magic method would be better (certainly prettier) long term solutions.