fix compatibility with PHP 5
This commit removes the dependency on the doctrine/persistence package
which requires PHP 7. The only class from that package the bundle relies
on is the ManagerRegistry.

On PHP 5, this class is provided by the doctrine/common package which
will be pulled in by doctrine/orm.

On PHP 7, the now needed doctrine/persistence package is installed as
doctrine/common (still required by doctrine/orm) depends on it.