Support for "required" tag attribute
I'm not sure that this is the right solution, but autocomplete fields do not have their "required" attribute set when the entity field is not nullable (input label is flagged with the red-dot, but form validation on submit ignores the input value). With this change it seems to work as intended.