Wrapped the action label (NOT icon) in a span
This allows for easier and/or more powerful styling of the action text and icon separately.

For example, here's my use case. In the index page, I want the icon to be visible always, but, when the screen width is below Bootstrap's 
"small" breakpoint (that is, when each entity gets its own "table" and the actions get their own row) I want the labels to become visible.

Without this, I'm unable to do that (or at least it becomes very cumbersome), because I can't target the label text not including the icon.

With this improvement, I can.