Update flash_messages.html.twig
The Flash messages add in the controller with `$this->addFlash()` did not always appear directly. 

While read the Symfony documentation, I noticed Symfony admin used the previous way to fetch and display the flash messages. Update it to latest version in documentation did fix the issue for us in our project.