Fix XSS in layout h1
the default layout does a `| raw` which can be exploited if the entity of your CRUD has a `__string()` method that uses some fields from the database.

example to reproduce. 


create an entity Article with a field 'title' and a method __string() which returns it. 

Then put as title `<script>alert('toto')</script>`  -> an XSS is triggered