Fix nullable detection
In Symfony production environment, since Doctrine 3.x, values for 'nullable' property can be either null, false or true. In Doctrine 2.x, it was only false or true. When Doctrine returned 'null' for the nullable property, the required option for the field was set to false.

While in fact, Doctrine considers 'nullable => null' as 'nullable => false', thus should make the field required.